Minor cleanups.

+++ src/client/talkmacros.cpp
@@ -1,7 +1,5 @@
 /*
- * XPilot NG CPP, a multiplayer space war game.
- *
- * Copyright (C) 1991-2001 by
+ * XPilot, a multiplayer gravity war game.  Copyright (C) 1991-2001 by
  *
  *      Bjørn Stabell
  *      Ken Ronny Schouten
@@ -204,7 +202,7 @@ static int Talk_macro_parse_mesg(char *outbuf, char *inbuf, long pos,
                 break;
             case 's':
                 if (pos < max - 1 - 6) /* short - "-16535" max no of chars */
-                    pos += sprintf(outbuf + pos, "%.2f", player->score);
+                    pos += sprintf(outbuf + pos, "%d", (int)player->score);
                 break;
             case 't':
                 if (BIT(Setup->mode, TEAM_PLAY))
@@ -357,6 +355,7 @@ static int Talk_macro_parse_mesg(char *outbuf, char *inbuf, long pos,
                     break;
                 case 'n':
                     outbuf[pos] = '\0';
+
                     if (Net_talk(outbuf) == -1)
                         return -1;
                     pos = 0;
